impl generating dbml file from simple dbml sample

+++ docker_clean.sh
@@ -1,4 +1,7 @@
-docker kill $(cat app_name) || echo '' > /dev/null
-docker rm $(cat app_name) && \
-docker rmi $(cat app_name) || \
+if [ "$(cat is_persistent)" = "Yes" ]; then
+  docker kill $(cat app_name) || echo '' > /dev/null
+  docker rm $(cat app_name)
+fi
+
+docker rmi lan22h/$(cat app_name):latest || \
 echo 'Nothing to clean!'
